Move useUpdatesState hook to expo-updates
