+++ packages/@expo/use-updates/docs/modules.md
@@ -17,7 +17,6 @@
 - [readLogEntries](modules.md#readlogentries)
 - [runUpdate](modules.md#runupdate)
 - [useUpdates](modules.md#useupdates)
-- [useUpdatesState](modules.md#useupdatesstate)
 
 ## Type Aliases
 
@@ -38,7 +37,7 @@ or an [`UpdateEvent`](#updateevent) emitted by native code.
 
 #### Defined in
 
-[UseUpdates.types.ts:69](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.types.ts#L69)
+[UseUpdates.types.ts:69](https://github.com/expo/expo/blob/f6be8ede73/packages/@expo/use-updates/src/UseUpdates.types.ts#L69)
 
 ___
 
@@ -63,7 +62,7 @@ Structure encapsulating information on the currently running app
 
 #### Defined in
 
-[UseUpdates.types.ts:15](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.types.ts#L15)
+[UseUpdates.types.ts:15](https://github.com/expo/expo/blob/f6be8ede73/packages/@expo/use-updates/src/UseUpdates.types.ts#L15)
 
 ___
 
@@ -87,7 +86,7 @@ The structures and methods returned by `useUpdates()`.
 
 #### Defined in
 
-[UseUpdates.types.ts:89](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.types.ts#L89)
+[UseUpdates.types.ts:89](https://github.com/expo/expo/blob/f6be8ede73/packages/@expo/use-updates/src/UseUpdates.types.ts#L89)
 
 ## Functions
 
@@ -105,7 +104,7 @@ If an error occurs, the `error` property will be set.
 
 #### Defined in
 
-[UseUpdates.ts:13](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.ts#L13)
+[UseUpdates.ts:13](https://github.com/expo/expo/blob/f6be8ede73/packages/@expo/use-updates/src/UseUpdates.ts#L13)
 
 ___
 
@@ -124,7 +123,7 @@ If an error occurs, the `error` property will be set.
 
 #### Defined in
 
-[UseUpdates.ts:40](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.ts#L40)
+[UseUpdates.ts:40](https://github.com/expo/expo/blob/f6be8ede73/packages/@expo/use-updates/src/UseUpdates.ts#L40)
 
 ___
 
@@ -147,7 +146,7 @@ If an error occurs, the `error` property will be set.
 
 #### Defined in
 
-[UseUpdates.ts:81](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.ts#L81)
+[UseUpdates.ts:81](https://github.com/expo/expo/blob/f6be8ede73/packages/@expo/use-updates/src/UseUpdates.ts#L81)
 
 ___
 
@@ -168,7 +167,7 @@ If an error occurs, the `error` property will be set.
 
 #### Defined in
 
-[UseUpdates.ts:66](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.ts#L66)
+[UseUpdates.ts:66](https://github.com/expo/expo/blob/f6be8ede73/packages/@expo/use-updates/src/UseUpdates.ts#L66)
 
 ___
 
@@ -233,25 +232,4 @@ the structures with information on currently running and available updates.
 
 #### Defined in
 
-[UseUpdates.ts:148](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.ts#L148)
-
-___
-
-### useUpdatesState
-
-▸ **useUpdatesState**(): `Object`
-
-Experimental hook to return the Updates state machine context maintained
-in native code.
-
-Eventually, this will be used to construct the information returned by `useUpdates()`.
-
-#### Returns
-
-`Object`
-
-A map of the state machine context.
-
-#### Defined in
-
-[UseUpdates.ts:218](https://github.com/expo/expo/blob/8f1b11dc48/packages/@expo/use-updates/src/UseUpdates.ts#L218)
+[UseUpdates.ts:148](https://github.com/expo/expo/blob/f6be8ede73/packages/@expo/use-updates/src/UseUpdates.ts#L148)
