+++ packages/@expo/use-updates/src/UseUpdatesEmitter.ts
@@ -15,10 +15,6 @@ function _getEmitter(): EventEmitter {
   if (!_emitter) {
     _emitter = new EventEmitter();
     DeviceEventEmitter.addListener('Expo.nativeUpdatesEvent', _emitNativeEvent);
-    DeviceEventEmitter.addListener(
-      'Expo.nativeUpdatesStateChangeEvent',
-      _emitNativeStateChangeEvent
-    );
   }
   return _emitter;
 }
@@ -55,26 +51,6 @@ function _emitNativeEvent(params: any) {
   _emitter.emit('Expo.useUpdatesEvent', newParams);
 }
 
-// Handle native state change events
-function _emitNativeStateChangeEvent(params: any) {
-  let newParams = { ...params };
-  if (typeof params === 'string') {
-    newParams = JSON.parse(params);
-  }
-  if (newParams.context.latestManifestString) {
-    newParams.context.latestManifest = JSON.parse(newParams.context.latestManifestString);
-    delete newParams.context.latestManifestString;
-  }
-  if (newParams.context.downloadedManifestString) {
-    newParams.context.downloadedManifest = JSON.parse(newParams.context.downloadedManifestString);
-    delete newParams.context.downloadedManifestString;
-  }
-  if (!_emitter) {
-    throw new Error(`EventEmitter must be initialized to use from its listener`);
-  }
-  _emitter?.emit('Expo.updatesStateChangeEvent', newParams);
-}
-
 // What JS code uses to emit events
 export const emitEvent = (event: UseUpdatesEvent) => {
   if (!_emitter) {
@@ -99,9 +75,4 @@ export const useUpdateEvents = (listener: (event: UseUpdatesEvent) => void) => {
     }
     return undefined;
   }, []);
-};
-
-export const addUpdatesStateChangeListener = (listener: (event: any) => void) => {
-  const emitter = _getEmitter();
-  return emitter.addListener('Expo.updatesStateChangeEvent', listener);
 };
