+++ packages/expo-updates/src/Updates.ts
@@ -15,6 +15,8 @@ import {
   UpdateFetchResult,
   UpdatesCheckAutomaticallyValue,
   UpdatesLogEntry,
+  UpdatesNativeStateChangeEvent,
+  UpdatesNativeStateMachineContext,
 } from './Updates.types';
 
 export * from './Updates.types';
@@ -303,6 +305,10 @@ function _getEmitter(): EventEmitter {
   if (!_emitter) {
     _emitter = new EventEmitter();
     DeviceEventEmitter.addListener('Expo.nativeUpdatesEvent', _emitEvent);
+    DeviceEventEmitter.addListener(
+      'Expo.nativeUpdatesStateChangeEvent',
+      _emitNativeStateChangeEvent
+    );
   }
   return _emitter;
 }
@@ -323,6 +329,26 @@ function _emitEvent(params): void {
   _emitter.emit('Expo.updatesEvent', newParams);
 }
 
+// Handle native state change events
+function _emitNativeStateChangeEvent(params: any) {
+  let newParams = { ...params };
+  if (typeof params === 'string') {
+    newParams = JSON.parse(params);
+  }
+  if (newParams.context.latestManifestString) {
+    newParams.context.latestManifest = JSON.parse(newParams.context.latestManifestString);
+    delete newParams.context.latestManifestString;
+  }
+  if (newParams.context.downloadedManifestString) {
+    newParams.context.downloadedManifest = JSON.parse(newParams.context.downloadedManifestString);
+    delete newParams.context.downloadedManifestString;
+  }
+  if (!_emitter) {
+    throw new Error(`EventEmitter must be initialized to use from its listener`);
+  }
+  _emitter?.emit('Expo.updatesStateChangeEvent', newParams);
+}
+
 /**
  * Adds a callback to be invoked when updates-related events occur (such as upon the initial app
  * load) due to auto-update settings chosen at build-time. See also the
@@ -341,7 +367,18 @@ export function addListener(listener: (event: UpdateEvent) => void): EventSubscr
 /**
  * @hidden
  */
-export async function nativeStateMachineContext(): Promise<{ [key: string]: any }> {
+export const addUpdatesStateChangeListener = (
+  listener: (event: UpdatesNativeStateChangeEvent) => void
+) => {
+  // Add listener for state change events
+  const emitter = _getEmitter();
+  return emitter.addListener('Expo.updatesStateChangeEvent', listener);
+};
+
+/**
+ * @hidden
+ */
+export async function nativeStateMachineContext(): Promise<UpdatesNativeStateMachineContext> {
   // Return the current state machine context
   if (!ExpoUpdates.nativeStateMachineContext) {
     throw new UnavailabilityError('Updates', 'readLogEntriesAsync');
